mvoe col visitor to mod rs as a common func
